feat: added shadcn buttons (#79)
