+++ client/src/app/components/button/Button.tsx
@@ -1,89 +1,54 @@
-'use client'
-import React, { FC } from 'react'
-import { twMerge } from 'tailwind-merge'
-import { type ClassValue, clsx } from 'clsx'
+import * as React from 'react'
+import { Slot } from '@radix-ui/react-slot'
+import { cva, type VariantProps } from 'class-variance-authority'
 
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
+import { cn } from '@/lib/utils'
+
+const buttonVariants = cva(
+  'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
+  {
+    variants: {
+      variant: {
+        default: 'bg-primary text-white hover:bg-primary/90',
+        destructive: 'bg-error text-white hover:bg-error/70',
+        outline:
+          'border border-primary text-primary bg-secondary/20 hover:bg-secondary/50 hover:text-accent',
+        secondary: 'bg-secondary text-primary hover:bg-secondary/80',
+        ghost: 'text-ghost-dark hover:text-ghost-light',
+        link: 'text-primary underline-offset-4 hover:underline',
+      },
+      size: {
+        default: 'h-10 px-4 py-2',
+        sm: 'h-9 rounded-md px-3',
+        lg: 'h-11 rounded-md px-8',
+        icon: 'h-10 w-10',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+      size: 'default',
+    },
+  },
+)
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean
 }
 
-type ButtonProps = {
-  theme:
-    | 'primary'
-    | 'secondary'
-    | 'ghost'
-    | 'warning'
-    | 'success'
-    | 'error'
-    | 'primary-out'
-    | 'secondary-out'
-  disabled?: boolean
-  click?: () => void
-  border: boolean
-}
-
-const themes: Record<
-  ButtonProps['theme'],
-  { bg: string; text: string; border?: string }
-> = {
-  primary: {
-    bg: 'bg-primary/50 hover:bg-primary/80',
-    text: 'text-primary-dark hover:text-white',
-    border: 'border-primary-dark/70',
-  },
-  secondary: {
-    bg: 'bg-secondary/60 hover:bg-secondary/80',
-    text: 'text-accent hover:text-white',
-    border: 'border-secondary',
-  },
-  warning: {
-    bg: 'bg-warning/10 hover:bg-warning/0',
-    text: 'text-warning',
-    border: 'border-warning hover:border-warning/50',
-  },
-  ghost: {
-    bg: 'bg-ghost-dark/10 hover:bg-ghost-dark/0',
-    text: 'text-ghost-dark',
-    border: 'border-ghost-light hover:border-ghost-light/50',
-  },
-  success: { bg: '', text: 'text-success', border: 'border-success' },
-  error: {
-    bg: 'bg-error/10 hover:bg-error/0',
-    text: 'text-error',
-    border: 'border-error hover:border-error/50',
-  },
-  'primary-out': {
-    bg: 'bg-primary/10 hover:bg-primary/0',
-    text: 'text-primary-dark',
-    border: 'border-primary-dark hover:border-primary-dark/50',
-  },
-  'secondary-out': {
-    bg: 'bg-secondary/10 hover:bg-secondary/0',
-    text: 'text-primary',
-    border: 'border-primary hover:border-primary/50',
-  },
-}
-
-const Button: FC<ButtonProps> = ({ theme, border, click }: ButtonProps) => {
-  const styles = themes[theme]
-  const className = React.useMemo(() => {
-    return cn(
-      `font-geist px-4 py-2 rounded-md font-semibold duration-300 cursor-pointer ${styles.bg} ${styles.text} ${border && `${styles.border} border-2`}`,
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, asChild = false, ...props }, ref) => {
+    const Comp = asChild ? Slot : 'button'
+    return (
+      <Comp
+        className={cn(buttonVariants({ variant, size, className }))}
+        ref={ref}
+        {...props}
+      />
     )
-  }, [theme, border])
-  return (
-    <button
-      className={className}
-      onClick={() => {
-        if (typeof click === 'function') {
-          click()
-        }
-      }}
-    >
-      <p>Button</p>
-    </button>
-  )
-}
+  },
+)
+Button.displayName = 'Button'
 
-const ButtonComponent = React.memo(Button)
-export default ButtonComponent
+export { Button, buttonVariants }
